update code refactor
