+++ frontend/src/routes/ProtectedPage.tsx.tsx
